Download UN DESA WPP data

Downloads data on demographic indicators in UN DESA WPP. Requires a working internet connection.

get_wpp(
  indicator = NULL,
  indicator_file_group = NULL,
  variant_id = 2,
  wpp_version = 2019,
  clean_names = FALSE,
  fct_age = TRUE,
  drop_id_cols = FALSE,
  tidy_pop_sex = FALSE,
  add_regions = FALSE,
  messages = TRUE
)

Arguments

indicator
Character string based on the name column in the wpp_indicators data frame. Represents the variables to be downloaded.

indicator_file_group
Character string based on the file_group column in the wpp_indicators data frame . Represents the file group to download data from. Only needed for obtaining different granularities of population data.

variant_id
Numeric value(s) based on the var_id column in the wpp_indicators data frame. Note, past data is in the "Medium" (2) variant only.

wpp_version
Integer for WPP version. Default of 2019. All WPP back to 1998 are available.

clean_names
Logical to indicate if column names should be cleaned

fct_age
Logical to indicate if AgeGrp column be converted to a factor.

drop_id_cols
Logical to indicate if VarId, LocID, MidPeriod, AgeGrpStart, AgeGrpSpan and SexID columns to be removed.

tidy_pop_sex
Logical to indicate if columns for sex specific population data should be stacked into single population column with an accompanying new sex column.

add_regions
Logical to indicate if to add a reg_name and area_name columns for countries (where LocID is less than 900)

messages
Logical to not suppress printing of messages.

Value

A tibble with downloaded data in tidy format

Details

Indicators must use the name corresponding to the name column in in the wpp_indicators data frame. The find_indicator() function can be used to look up the indicator code and availability by variants There are 35 different indicators in WPP data (starting from 1998)

name	details	file_group

ASFR	Age-specific fertility rate (births per 1,000 women)	Fertility_by_Age

ax	Average number of years lived (nax) between ages x and x+n by those dying in the interval	Life_Table

Births	Number of births, both sexes combined (thousands)	Fertility_by_Age

Births	Number of births, both sexes combined (thousands)	Period_Indicators

CBR	Crude birth rate (births per 1,000 population)	Period_Indicators

CDR	Crude death rate (deaths per 1,000 population)	Period_Indicators

CNMR	Net migration rate (per 1,000 population)	Period_Indicators

Deaths	Number of deaths, both sexes combined (thousands)	Period_Indicators

DeathsFemale	Number of female deaths (thousands)	Period_Indicators

DeathsMale	Number of male deaths (thousands)	Period_Indicators

dx	Number of deaths, (ndx), between ages x and x+n	Life_Table

ex	Expectation of life (ex) at age x, i.e., average number of years lived subsequent to age x by those reaching age x	Life_Table

GrowthRate	Average annual rate of population change (percentage)	Period_Indicators

IMR	Infant mortality rate, q(1), for both sexes combined (infant deaths per 1,000 live births)	Period_Indicators

LEx	Life expectancy at birth for both sexes combined (years)	Period_Indicators

LExFemale	Female life expectancy at birth (years)	Period_Indicators

LExMale	Male life expectancy at birth (years)	Period_Indicators

lx	Number of survivors, (lx), at age (x) for 100000 births	Life_Table

Lx	Number of person-years lived, (nLx), between ages x and x+n	Life_Table

MAC	Female mean age of childbearing (years)	Period_Indicators

mx	Central death rate, nmx, for the age interval (x, x+n)	Life_Table

NatIncr	Rate of natural increase (per 1,000 population)	Period_Indicators

NetMigrations	Net number of migrants, both sexes combined (thousands)	Period_Indicators

NRR	Net reproduction rate (surviving daughters per woman)	Period_Indicators

PASFR	Percentage age-specific fertility rate	Fertility_by_Age

PopDensity	Population per square kilometre (thousands)	TotalPopulationBySex

PopFemale	Female population in the age group (thousands)	PopulationByAgeSex

PopFemale	Total female population (thousands)	TotalPopulationBySex

PopFemale	Female population for the individual age (thousands)	PopulationBySingleAgeSex

PopFemale	Female population in the age group (thousands)	PopulationByAgeSex_5x5

PopMale	Male population in the age group (thousands)	PopulationByAgeSex

PopMale	Total male population (thousands)	TotalPopulationBySex

PopMale	Male population for the individual age (thousands)	PopulationBySingleAgeSex

PopMale	Male population in the age group (thousands)	PopulationByAgeSex_5x5

PopTotal	Total population in the age group (thousands)	PopulationByAgeSex

PopTotal	Total population, both sexes (thousands)	TotalPopulationBySex

PopTotal	Total population for the individual age (thousands)	PopulationBySingleAgeSex

PopTotal	Total population in the age group (thousands)	PopulationByAgeSex_5x5

px	Probability of surviving, (npx), for an individual of age x to age x+n	Life_Table

Q5	Under-five mortality, 5q0, for both sexes combined (deaths under age five per 1,000 live births)	Period_Indicators

qx	Probability of dying (nqx), for an individual between age x and x+n	Life_Table

SRB	Sex ratio at birth (male births per female births)	Period_Indicators

Sx	Survival ratio (nSx) corresponding to proportion of the life table population in age group (x, x+n) who are alive n year later	Life_Table

TFR	Total fertility (live births per woman)	Period_Indicators

Tx	Person-years lived, (Tx), above age x	Life_Table

The variant_id argument must be one or more numbers from the var_id column in the wpp_indicators data frame. Not all indicators area available in all variants. Use the find_indicator() function to check availability. There are 14 different variants in WPP data (starting from 1998).

var_id	variant

2	Medium

3	High

4	Low

5	Constant fertility

6	Instant replacement

7	Zero migration

8	Constant mortality

9	No change

10	Momentum

202	Median PI (BHM median in WPP2015)

203	Upper 80 PI

204	Lower 80 PI

206	Upper 95 PI

207	Lower 95 PI

Examples

# \donttest{
# single indicator from medium variant of latest WPP
get_wpp(indicator = "TFR")

#> # A tibble: 14,940 × 7
#>    LocID Location    VarID Variant Time      MidPeriod   TFR
#>    <dbl> <chr>       <dbl> <chr>   <chr>         <dbl> <dbl>
#>  1     4 Afghanistan     2 Medium  1950-1955      1953  7.45
#>  2     4 Afghanistan     2 Medium  1955-1960      1958  7.45
#>  3     4 Afghanistan     2 Medium  1960-1965      1963  7.45
#>  4     4 Afghanistan     2 Medium  1965-1970      1968  7.45
#>  5     4 Afghanistan     2 Medium  1970-1975      1973  7.45
#>  6     4 Afghanistan     2 Medium  1975-1980      1978  7.45
#>  7     4 Afghanistan     2 Medium  1980-1985      1983  7.45
#>  8     4 Afghanistan     2 Medium  1985-1990      1988  7.47
#>  9     4 Afghanistan     2 Medium  1990-1995      1993  7.48
#> 10     4 Afghanistan     2 Medium  1995-2000      1998  7.65
#> # … with 14,930 more rows

# single indicator from multiple variants of latest WPP
get_wpp(indicator = "TFR", variant_id = c(2, 3, 4))

#> # A tibble: 24,060 × 7
#>    LocID Location    VarID Variant Time      MidPeriod   TFR
#>    <dbl> <chr>       <dbl> <chr>   <chr>         <dbl> <dbl>
#>  1     4 Afghanistan     2 Medium  1950-1955      1953  7.45
#>  2     4 Afghanistan     2 Medium  1955-1960      1958  7.45
#>  3     4 Afghanistan     2 Medium  1960-1965      1963  7.45
#>  4     4 Afghanistan     2 Medium  1965-1970      1968  7.45
#>  5     4 Afghanistan     2 Medium  1970-1975      1973  7.45
#>  6     4 Afghanistan     2 Medium  1975-1980      1978  7.45
#>  7     4 Afghanistan     2 Medium  1980-1985      1983  7.45
#>  8     4 Afghanistan     2 Medium  1985-1990      1988  7.47
#>  9     4 Afghanistan     2 Medium  1990-1995      1993  7.48
#> 10     4 Afghanistan     2 Medium  1995-2000      1998  7.65
#> # … with 24,050 more rows

# multiple population indicators from single variant of latest WPP
get_wpp(indicator = c("PopTotal", "PopMale", "PopFemale"))

#> Downloading from PopulationByAgeSex

#> Also available in: TotalPopulationBySex

#> Also available in: PopulationBySingleAgeSex

#> Use indicator_file_group to get alternative measures

#> # A tibble: 1,404,753 × 12
#>    LocID Location    VarID Variant  Time MidPeriod AgeGrp AgeGrpStart AgeGrpSpan
#>    <dbl> <chr>       <dbl> <chr>   <dbl>     <dbl> <fct>        <dbl>      <dbl>
#>  1     4 Afghanistan     2 Medium   1950     1950. 0-4              0          5
#>  2     4 Afghanistan     2 Medium   1950     1950. 5-9              5          5
#>  3     4 Afghanistan     2 Medium   1950     1950. 10-14           10          5
#>  4     4 Afghanistan     2 Medium   1950     1950. 15-19           15          5
#>  5     4 Afghanistan     2 Medium   1950     1950. 20-24           20          5
#>  6     4 Afghanistan     2 Medium   1950     1950. 25-29           25          5
#>  7     4 Afghanistan     2 Medium   1950     1950. 30-34           30          5
#>  8     4 Afghanistan     2 Medium   1950     1950. 35-39           35          5
#>  9     4 Afghanistan     2 Medium   1950     1950. 40-44           40          5
#> 10     4 Afghanistan     2 Medium   1950     1950. 45-49           45          5
#> # … with 1,404,743 more rows, and 3 more variables: PopFemale <dbl>,
#> #   PopMale <dbl>, PopTotal <dbl>

# there are multiple population indicators in the WPP with different levels of granularity
# use indicator_file_group to select the desired version of population indicator(s)
get_wpp(indicator = c("PopTotal", "PopMale", "PopFemale"), indicator_file_group =  "TotalPopulationBySex")

#> # A tibble: 72,027 × 9
#>    LocID Location    VarID Variant  Time MidPeriod PopFemale PopMale PopTotal
#>    <dbl> <chr>       <dbl> <chr>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1     4 Afghanistan     2 Medium   1950     1950.     3653.   4099.    7752.
#>  2     4 Afghanistan     2 Medium   1951     1952.     3705.   4135.    7840.
#>  3     4 Afghanistan     2 Medium   1952     1952.     3762.   4174.    7936.
#>  4     4 Afghanistan     2 Medium   1953     1954.     3821.   4218.    8040.
#>  5     4 Afghanistan     2 Medium   1954     1954.     3885.   4266.    8151.
#>  6     4 Afghanistan     2 Medium   1955     1956.     3952.   4319.    8271.
#>  7     4 Afghanistan     2 Medium   1956     1956.     4023.   4376.    8399.
#>  8     4 Afghanistan     2 Medium   1957     1958.     4098    4437.    8535.
#>  9     4 Afghanistan     2 Medium   1958     1958.     4177.   4503.    8680.
#> 10     4 Afghanistan     2 Medium   1959     1960.     4260.   4574.    8834.
#> # … with 72,017 more rows

# tidy sex into a single column and drop id columns
get_wpp(indicator = c("PopMale", "PopFemale"), indicator_file_group =  "TotalPopulationBySex",
        tidy_pop_sex = TRUE, drop_id_cols = TRUE)

#> # A tibble: 144,054 × 5
#>    Location    Variant  Time Sex      Pop
#>    <chr>       <chr>   <dbl> <chr>  <dbl>
#>  1 Afghanistan Medium   1950 Female 3653.
#>  2 Afghanistan Medium   1950 Male   4099.
#>  3 Afghanistan Medium   1951 Female 3705.
#>  4 Afghanistan Medium   1951 Male   4135.
#>  5 Afghanistan Medium   1952 Female 3762.
#>  6 Afghanistan Medium   1952 Male   4174.
#>  7 Afghanistan Medium   1953 Female 3821.
#>  8 Afghanistan Medium   1953 Male   4218.
#>  9 Afghanistan Medium   1954 Female 3885.
#> 10 Afghanistan Medium   1954 Male   4266.
#> # … with 144,044 more rows

# clean column names
get_wpp(indicator = c("SRB", "NetMigrations", "GrowthRate"), clean_names = TRUE, drop_id_cols = TRUE)

#> # A tibble: 14,940 × 6
#>    location    variant time      growth_rate net_migrations   srb
#>    <chr>       <chr>   <chr>           <dbl>          <dbl> <dbl>
#>  1 Afghanistan Medium  1950-1955       1.30            -20   1.06
#>  2 Afghanistan Medium  1955-1960       1.68            -20   1.06
#>  3 Afghanistan Medium  1960-1965       2.03            -20   1.06
#>  4 Afghanistan Medium  1965-1970       2.31            -20   1.06
#>  5 Afghanistan Medium  1970-1975       2.54            -20   1.06
#>  6 Afghanistan Medium  1975-1980       1.02          -1154.  1.06
#>  7 Afghanistan Medium  1980-1985      -2.24          -3345.  1.06
#>  8 Afghanistan Medium  1985-1990       0.779         -1525.  1.06
#>  9 Afghanistan Medium  1990-1995       7.56           3076.  1.06
#> 10 Afghanistan Medium  1995-2000       2.75           -868.  1.06
#> # … with 14,930 more rows

# }