Find available indicator code names or file groups for UN DESA World Population Prospects data

Find available indicator code names or file groups for UN DESA World Population Prospects data

Usage

find_indicator(x, wpp_version = 2022, simple = TRUE)

Arguments

x

Character string on key word or name related to indicator of potential interest.

wpp_version

Integer for WPP version. Default of 2022. All WPP back to 1998 are available.

simple

Logical to give simple table without variant or file group information

Value

A subset of the wpp_indicators data frame with one or more of columns matching the keyword given to x. Use the result in the indicator column to input to the get_wpp() function for downloading data.

Examples

find_indicator("migration")
#> # A tibble: 2 × 3
#>   name          details                file                  
#>   <chr>         <chr>                  <chr>                 
#> 1 NetMigrations Net Number of Migrants Demographic_Indicators
#> 2 CNMR          Net Migration Rate     Demographic_Indicators

find_indicator("sex ratio")
#> # A tibble: 2 × 3
#>   name        details                            file                  
#>   <chr>       <chr>                              <chr>                 
#> 1 PopSexRatio Population Sex Ratio, as of 1 July Demographic_Indicators
#> 2 SRB         Sex Ratio at Birth                 Demographic_Indicators

# show variant information
find_indicator("Deaths", simple = FALSE)
#> # A tibble: 62 × 10
#>    name        details unit  var_id variant   wpp topic file  file0 file_details
#>    <chr>       <chr>   <chr>  <dbl> <chr>   <dbl> <chr> <chr> <chr> <chr>       
#>  1 NatChange   Natura… thou…     10 Moment…  2022 Popu… Demo… Demo… Demographic…
#>  2 Deaths      Total … thou…     10 Moment…  2022 Mort… Demo… Demo… Demographic…
#>  3 DeathsMale  Male D… thou…     10 Moment…  2022 Mort… Demo… Demo… Demographic…
#>  4 DeathsFema… Female… thou…     10 Moment…  2022 Mort… Demo… Demo… Demographic…
#>  5 InfantDeat… Infant… thou…     10 Moment…  2022 Mort… Demo… Demo… Demographic…
#>  6 Under5Deat… Deaths… thou…     10 Moment…  2022 Mort… Demo… Demo… Demographic…
#>  7 NatChange   Natura… thou…     16 Instan…  2022 Popu… Demo… Demo… Demographic…
#>  8 Deaths      Total … thou…     16 Instan…  2022 Mort… Demo… Demo… Demographic…
#>  9 DeathsMale  Male D… thou…     16 Instan…  2022 Mort… Demo… Demo… Demographic…
#> 10 DeathsFema… Female… thou…     16 Instan…  2022 Mort… Demo… Demo… Demographic…
#> # … with 52 more rows