Download UN DESA WPP data

Downloads data on demographic indicators in UN DESA WPP. Requires a working internet connection.

get_wpp(
  indicator = NULL,
  indicator_file_group = NULL,
  variant_id = 2,
  wpp_version = 2019,
  clean_names = FALSE,
  fct_age = TRUE,
  drop_id_cols = FALSE,
  tidy_pop_sex = FALSE,
  add_regions = FALSE,
  messages = TRUE
)

Arguments

indicator
Character string based on the name column in the wpp_indicators data frame. Represents the variables to be downloaded.

indicator_file_group
Character string based on the file_group column in the wpp_indicators data frame . Represents the file group to download data from. Only needed for obtaining different granularities of population data.

variant_id
Numeric value(s) based on the var_id column in the wpp_indicators data frame. Note, past data is in the "Medium" (2) variant only.

wpp_version
Integer for WPP version. Default of 2019. All WPP back to 1998 are available.

clean_names
Logical to indicate if column names should be cleaned

fct_age
Logical to indicate if AgeGrp column be converted to a factor.

drop_id_cols
Logical to indicate if VarId, LocID, MidPeriod, AgeGrpStart, AgeGrpSpan and SexID columns to be removed.

tidy_pop_sex
Logical to indicate if columns for sex specific population data should be stacked into single population column with an accompanying new sex column.

add_regions
Logical to indicate if to add a reg_name and area_name columns for countries (where LocID is less than 900)

messages
Logical to not suppress printing of messages.

Value

A tibble with downloaded data in tidy format

Details

Indicators must use the name corresponding to the name column in in the wpp_indicators data frame. The find_indicator() function can be used to look up the indicator code and availability by variants There are 35 different indicators in WPP data (starting from 1998)

name	details	file_group

ASFR	Age-specific fertility rate (births per 1,000 women)	Fertility_by_Age

ax	Average number of years lived (nax) between ages x and x+n by those dying in the interval	Life_Table

Births	Number of births, both sexes combined (thousands)	Fertility_by_Age

Births	Number of births, both sexes combined (thousands)	Period_Indicators

CBR	Crude birth rate (births per 1,000 population)	Period_Indicators

CDR	Crude death rate (deaths per 1,000 population)	Period_Indicators

CNMR	Net migration rate (per 1,000 population)	Period_Indicators

Deaths	Number of deaths, both sexes combined (thousands)	Period_Indicators

DeathsFemale	Number of female deaths (thousands)	Period_Indicators

DeathsMale	Number of male deaths (thousands)	Period_Indicators

dx	Number of deaths, (ndx), between ages x and x+n	Life_Table

ex	Expectation of life (ex) at age x, i.e., average number of years lived subsequent to age x by those reaching age x	Life_Table

GrowthRate	Average annual rate of population change (percentage)	Period_Indicators

IMR	Infant mortality rate, q(1), for both sexes combined (infant deaths per 1,000 live births)	Period_Indicators

LEx	Life expectancy at birth for both sexes combined (years)	Period_Indicators

LExFemale	Female life expectancy at birth (years)	Period_Indicators

LExMale	Male life expectancy at birth (years)	Period_Indicators

lx	Number of survivors, (lx), at age (x) for 100000 births	Life_Table

Lx	Number of person-years lived, (nLx), between ages x and x+n	Life_Table

MAC	Female mean age of childbearing (years)	Period_Indicators

mx	Central death rate, nmx, for the age interval (x, x+n)	Life_Table

NatIncr	Rate of natural increase (per 1,000 population)	Period_Indicators

NetMigrations	Net number of migrants, both sexes combined (thousands)	Period_Indicators

NRR	Net reproduction rate (surviving daughters per woman)	Period_Indicators

PASFR	Percentage age-specific fertility rate	Fertility_by_Age

PopDensity	Population per square kilometre (thousands)	TotalPopulationBySex

PopFemale	Female population in the age group (thousands)	PopulationByAgeSex

PopFemale	Total female population (thousands)	TotalPopulationBySex

PopFemale	Female population for the individual age (thousands)	PopulationBySingleAgeSex

PopFemale	Female population in the age group (thousands)	PopulationByAgeSex_5x5

PopMale	Male population in the age group (thousands)	PopulationByAgeSex

PopMale	Total male population (thousands)	TotalPopulationBySex

PopMale	Male population for the individual age (thousands)	PopulationBySingleAgeSex

PopMale	Male population in the age group (thousands)	PopulationByAgeSex_5x5

PopTotal	Total population in the age group (thousands)	PopulationByAgeSex

PopTotal	Total population, both sexes (thousands)	TotalPopulationBySex

PopTotal	Total population for the individual age (thousands)	PopulationBySingleAgeSex

PopTotal	Total population in the age group (thousands)	PopulationByAgeSex_5x5

px	Probability of surviving, (npx), for an individual of age x to age x+n	Life_Table

Q5	Under-five mortality, 5q0, for both sexes combined (deaths under age five per 1,000 live births)	Period_Indicators

qx	Probability of dying (nqx), for an individual between age x and x+n	Life_Table

SRB	Sex ratio at birth (male births per female births)	Period_Indicators

Sx	Survival ratio (nSx) corresponding to proportion of the life table population in age group (x, x+n) who are alive n year later	Life_Table

TFR	Total fertility (live births per woman)	Period_Indicators

Tx	Person-years lived, (Tx), above age x	Life_Table

The variant_id argument must be one or more numbers from the var_id column in the wpp_indicators data frame. Not all indicators area available in all variants. Use the find_indicator() function to check availability. There are 14 different variants in WPP data (starting from 1998).

var_id	variant

2	Medium

3	High

4	Low

5	Constant fertility

6	Instant replacement

7	Zero migration

8	Constant mortality

9	No change

10	Momentum

202	Median PI (BHM median in WPP2015)

203	Upper 80 PI

204	Lower 80 PI

206	Upper 95 PI

207	Lower 95 PI

Examples

# \donttest{
# single indicator from medium variant of latest WPP
get_wpp(indicator = "TFR")

#> Error: Problem with `mutate()` column `i`.
#> ℹ `i = purrr::map(...)`.
#> ✖ there is no package called ‘vroom’

# single indicator from multiple variants of latest WPP
get_wpp(indicator = "TFR", variant_id = c(2, 3, 4))

#> Error: Problem with `mutate()` column `i`.
#> ℹ `i = purrr::map(...)`.
#> ✖ there is no package called ‘vroom’

# multiple population indicators from single variant of latest WPP
get_wpp(indicator = c("PopTotal", "PopMale", "PopFemale"))

#> Downloading from PopulationByAgeSex

#> Also available in: TotalPopulationBySex

#> Also available in: PopulationBySingleAgeSex

#> Use indicator_file_group to get alternative measures

#> Error: Problem with `mutate()` column `i`.
#> ℹ `i = purrr::map(...)`.
#> ✖ there is no package called ‘vroom’

# there are multiple population indicators in the WPP with different levels of granularity
# use indicator_file_group to select the desired version of population indicator(s)
get_wpp(indicator = c("PopTotal", "PopMale", "PopFemale"), indicator_file_group =  "TotalPopulationBySex")

#> Error: Problem with `mutate()` column `i`.
#> ℹ `i = purrr::map(...)`.
#> ✖ there is no package called ‘vroom’

# tidy sex into a single column and drop id columns
get_wpp(indicator = c("PopMale", "PopFemale"), indicator_file_group =  "TotalPopulationBySex",
        tidy_pop_sex = TRUE, drop_id_cols = TRUE)

#> Error: Problem with `mutate()` column `i`.
#> ℹ `i = purrr::map(...)`.
#> ✖ there is no package called ‘vroom’

# clean column names
get_wpp(indicator = c("SRB", "NetMigrations", "GrowthRate"), clean_names = TRUE, drop_id_cols = TRUE)

#> Error: Problem with `mutate()` column `i`.
#> ℹ `i = purrr::map(...)`.
#> ✖ there is no package called ‘vroom’

# }