Download UN DESA WPP data

Downloads data on demographic indicators in UN DESA WPP. Requires a working internet connection.

get_wpp(
  indicator = NULL,
  indicator_file = NULL,
  pop_age = c("total", "single", "five"),
  pop_sex = c("total", "both", "male", "female", "all"),
  pop_freq = c("annual", "five"),
  pop_date = c("jul1", "jan1", "jan1-dec31"),
  variant_id = 2,
  wpp_version = 2022,
  clean_names = FALSE,
  fct_age = TRUE,
  drop_id_cols = FALSE,
  tidy_pop_sex = FALSE,
  add_regions = FALSE,
  add_iso_codes = FALSE,
  messages = TRUE,
  server = c("github", "local")
)

Arguments

indicator

Character string based on the name column in the wpp_indicators data frame or pop. Represents the variables to be downloaded.

indicator_file

Character string based on the file column in the wpp_indicators data frame . Represents the file group to download data from. Needed for obtaining indicators that are available with different levels of granularities (such as Births in overall population or Births by mothers age group).

pop_age

Character string for population age groups if indicator is set to pop. Defaults to no age groups total, but can be set to single or five.

pop_sex

Character string for population sexes if indicatoris set to pop. Defaults to no sex total, but can be set to male, female, both or all.

pop_freq

Character string for frequency of population data if indicator is set to pop. Defaults to annual, but in a some (exceptional cases) can be set to five.

pop_date

Character string for frequency of population data if indicator is set to pop. Defaults to jul1 (July 1st), but for WPP2022 can be set to jan1 for population at beginning of year or jan1-dec31 for exposure population.

variant_id

Numeric value(s) based on the var_id column in the wpp_indicators data frame. Note, past data is in the "Medium" (2) variant only.

wpp_version

Integer for WPP version. Default of 2019. All WPP back to 1998 are available.

clean_names

Logical to indicate if column names should be cleaned

fct_age

Logical to indicate if AgeGrp column be converted to a factor.

drop_id_cols

Logical to indicate if VarId, LocID, MidPeriod, AgeGrpStart, AgeGrpSpan and SexID columns to be removed.

tidy_pop_sex

Logical to indicate if columns for sex specific population data should be stacked into single population column with an accompanying new sex column.

add_regions

Logical to indicate if to add a reg_name and area_name columns for countries (where LocID is less than 900)

add_iso_codes

Logical to indicate if to add a iso3 and iso2 columns for ISO 3 and 2 letter country codes (where LocID is less than 900)

messages

Logical to not suppress printing of messages.

server

Character string for location to download data from. Default of github.

Value

A tibble with downloaded data in tidy format

Details

Indicators must use the name corresponding to the name column in in the wpp_indicators data frame. The find_indicator() function can be used to look up the indicator code and availability by variants

There are 114 different indicators in WPP data (starting from 1998). See the full table of the different indicators available in each WPP.

The variant_id argument must be one or more numbers from the var_id column in the wpp_indicators data frame. Not all indicators area available in all variants. Use the find_indicator() function to check availability. There are 14 different variants in WPP data (starting from 1998).

var_id	variant

2	Medium

3	High

4	Low

5	Constant fertility

6	Instant replacement

7	Zero migration

8	Constant mortality

9	No change

10	Momentum

16	Instant replacement zero migration

202	Median PI (BHM median in WPP2015)

203	Upper 80 PI

204	Lower 80 PI

206	Upper 95 PI

207	Lower 95 PI

Examples

# \donttest{
# single indicator from medium variant of latest WPP
get_wpp(indicator = "TFR")
#> # A tibble: 43,472 × 6
#>    LocID Location VarID Variant  Time   TFR
#>    <dbl> <chr>    <dbl> <chr>   <dbl> <dbl>
#>  1   900 World        2 Medium   1950  4.86
#>  2   900 World        2 Medium   1951  4.83
#>  3   900 World        2 Medium   1952  5.01
#>  4   900 World        2 Medium   1953  4.94
#>  5   900 World        2 Medium   1954  5.01
#>  6   900 World        2 Medium   1955  5.01
#>  7   900 World        2 Medium   1956  4.94
#>  8   900 World        2 Medium   1957  5.08
#>  9   900 World        2 Medium   1958  4.93
#> 10   900 World        2 Medium   1959  4.74
#> # … with 43,462 more rows

# single indicator from multiple variants of latest WPP
get_wpp(indicator = "TFR", variant_id = c(2, 3, 4))
#> # A tibble: 89,232 × 6
#>    LocID Location VarID Variant  Time   TFR
#>    <dbl> <chr>    <dbl> <chr>   <dbl> <dbl>
#>  1   900 World        2 Medium   1950  4.86
#>  2   900 World        2 Medium   1951  4.83
#>  3   900 World        2 Medium   1952  5.01
#>  4   900 World        2 Medium   1953  4.94
#>  5   900 World        2 Medium   1954  5.01
#>  6   900 World        2 Medium   1955  5.01
#>  7   900 World        2 Medium   1956  4.94
#>  8   900 World        2 Medium   1957  5.08
#>  9   900 World        2 Medium   1958  4.93
#> 10   900 World        2 Medium   1959  4.74
#> # … with 89,222 more rows

# some indicators appear in multiple file groups, for example Births
# represents total number of births in the country in the
# Demographic_Indicators file (chosen by default)
get_wpp(indicator = "Births")
#> Indicator(s) appears in more than one file.
#> 
#> Only downloading indicator(s) from file: Demographic_Indicators 
#> 
#> Need multiple get_wpp() calls to get indicators in different files. See ?wpp_indicators and ?find_indicators for more information on files.
#> # A tibble: 43,472 × 6
#>    LocID Location VarID Variant  Time  Births
#>    <dbl> <chr>    <dbl> <chr>   <dbl>   <dbl>
#>  1   900 World        2 Medium   1950  92083.
#>  2   900 World        2 Medium   1951  92837.
#>  3   900 World        2 Medium   1952  97607.
#>  4   900 World        2 Medium   1953  97556.
#>  5   900 World        2 Medium   1954 100348.
#>  6   900 World        2 Medium   1955 101807.
#>  7   900 World        2 Medium   1956 101827.
#>  8   900 World        2 Medium   1957 105978.
#>  9   900 World        2 Medium   1958 104557.
#> 10   900 World        2 Medium   1959 101922.
#> # … with 43,462 more rows

# specify indicator_file to get number of births by mothers 5-year age group
get_wpp(indicator = c("Births", "ASFR"),
        indicator_file = "Fertility_by_Age5",
        drop_id_cols = TRUE)
#> # A tibble: 387,315 × 6
#>    Location Variant  Time AgeGrp    ASFR  Births
#>    <chr>    <chr>   <dbl> <fct>    <dbl>   <dbl>
#>  1 World    Medium   1950 10-14    4.54    571. 
#>  2 World    Medium   1950 15-19   91.1   10629. 
#>  3 World    Medium   1950 20-24  229.    24991. 
#>  4 World    Medium   1950 25-29  241.    23498. 
#>  5 World    Medium   1950 30-34  196.    16100. 
#>  6 World    Medium   1950 35-39  132.    10661. 
#>  7 World    Medium   1950 40-44   64.6    4703. 
#>  8 World    Medium   1950 45-49   14.0     888. 
#>  9 World    Medium   1950 50-54    0.755    40.4
#> 10 World    Medium   1951 10-14    4.64    587. 
#> # … with 387,305 more rows

# PopTotal, PopMale and PopFemale indicators are in many WPP files with
# a wide range granularity. Set indicator = "pop" and use the pop_sex,
# pop_age, pop_freq and pop_date to get desired data from the appropriate
# indicator_file...

# when using indicator = "pop" get_wpp() defaults to annual total population
# (summed over age and sex)
get_wpp(indicator = "pop")
#> Indicator(s) appears in more than one file.
#> 
#> Only downloading indicator(s) from file: Population1JanuaryByAge5GroupSex 
#> 
#> Need multiple get_wpp() calls to get indicators in different files. See ?wpp_indicators and ?find_indicators for more information on files.
#> # A tibble: 909,720 × 10
#>    LocID Location VarID Variant  Time MidPeriod AgeGrp AgeGrpStart AgeGrpSpan
#>    <dbl> <chr>    <dbl> <chr>   <dbl>     <dbl> <fct>        <dbl>      <dbl>
#>  1   900 World        2 Medium   1950      1950 0-4              0          5
#>  2   900 World        2 Medium   1950      1950 5-9              5          5
#>  3   900 World        2 Medium   1950      1950 10-14           10          5
#>  4   900 World        2 Medium   1950      1950 15-19           15          5
#>  5   900 World        2 Medium   1950      1950 20-24           20          5
#>  6   900 World        2 Medium   1950      1950 25-29           25          5
#>  7   900 World        2 Medium   1950      1950 30-34           30          5
#>  8   900 World        2 Medium   1950      1950 35-39           35          5
#>  9   900 World        2 Medium   1950      1950 40-44           40          5
#> 10   900 World        2 Medium   1950      1950 45-49           45          5
#> # … with 909,710 more rows, and 1 more variable: PopTotal <dbl>

# use pop_sex to get specific sexes (or both or all)
get_wpp(indicator = "pop", pop_sex = "male")
#> Indicator(s) appears in more than one file.
#> 
#> Only downloading indicator(s) from file: Population1JanuaryByAge5GroupSex 
#> 
#> Need multiple get_wpp() calls to get indicators in different files. See ?wpp_indicators and ?find_indicators for more information on files.
#> # A tibble: 909,720 × 10
#>    LocID Location VarID Variant  Time MidPeriod AgeGrp AgeGrpStart AgeGrpSpan
#>    <dbl> <chr>    <dbl> <chr>   <dbl>     <dbl> <fct>        <dbl>      <dbl>
#>  1   900 World        2 Medium   1950      1950 0-4              0          5
#>  2   900 World        2 Medium   1950      1950 5-9              5          5
#>  3   900 World        2 Medium   1950      1950 10-14           10          5
#>  4   900 World        2 Medium   1950      1950 15-19           15          5
#>  5   900 World        2 Medium   1950      1950 20-24           20          5
#>  6   900 World        2 Medium   1950      1950 25-29           25          5
#>  7   900 World        2 Medium   1950      1950 30-34           30          5
#>  8   900 World        2 Medium   1950      1950 35-39           35          5
#>  9   900 World        2 Medium   1950      1950 40-44           40          5
#> 10   900 World        2 Medium   1950      1950 45-49           45          5
#> # … with 909,710 more rows, and 1 more variable: PopMale <dbl>

# use pop_age to specify age groups
get_wpp(indicator = "pop", pop_sex = "both", pop_age = "five")
#> Indicator(s) appears in more than one file.
#> 
#> Only downloading indicator(s) from file: Population1JanuaryByAge5GroupSex 
#> 
#> Need multiple get_wpp() calls to get indicators in different files. See ?wpp_indicators and ?find_indicators for more information on files.
#> # A tibble: 909,720 × 11
#>    LocID Location VarID Variant  Time MidPeriod AgeGrp AgeGrpStart AgeGrpSpan
#>    <dbl> <chr>    <dbl> <chr>   <dbl>     <dbl> <fct>        <dbl>      <dbl>
#>  1   900 World        2 Medium   1950      1950 0-4              0          5
#>  2   900 World        2 Medium   1950      1950 5-9              5          5
#>  3   900 World        2 Medium   1950      1950 10-14           10          5
#>  4   900 World        2 Medium   1950      1950 15-19           15          5
#>  5   900 World        2 Medium   1950      1950 20-24           20          5
#>  6   900 World        2 Medium   1950      1950 25-29           25          5
#>  7   900 World        2 Medium   1950      1950 30-34           30          5
#>  8   900 World        2 Medium   1950      1950 35-39           35          5
#>  9   900 World        2 Medium   1950      1950 40-44           40          5
#> 10   900 World        2 Medium   1950      1950 45-49           45          5
#> # … with 909,710 more rows, and 2 more variables: PopMale <dbl>,
#> #   PopFemale <dbl>

# use pop_date to specify populations at start of year (rather than mid-year)
get_wpp(indicator = "pop", pop_sex = "female", pop_age = "five", pop_date = "jan1")
#> Indicator(s) appears in more than one file.
#> 
#> Only downloading indicator(s) from file: Population1JanuaryByAge5GroupSex 
#> 
#> Need multiple get_wpp() calls to get indicators in different files. See ?wpp_indicators and ?find_indicators for more information on files.
#> # A tibble: 909,720 × 10
#>    LocID Location VarID Variant  Time MidPeriod AgeGrp AgeGrpStart AgeGrpSpan
#>    <dbl> <chr>    <dbl> <chr>   <dbl>     <dbl> <fct>        <dbl>      <dbl>
#>  1   900 World        2 Medium   1950      1950 0-4              0          5
#>  2   900 World        2 Medium   1950      1950 5-9              5          5
#>  3   900 World        2 Medium   1950      1950 10-14           10          5
#>  4   900 World        2 Medium   1950      1950 15-19           15          5
#>  5   900 World        2 Medium   1950      1950 20-24           20          5
#>  6   900 World        2 Medium   1950      1950 25-29           25          5
#>  7   900 World        2 Medium   1950      1950 30-34           30          5
#>  8   900 World        2 Medium   1950      1950 35-39           35          5
#>  9   900 World        2 Medium   1950      1950 40-44           40          5
#> 10   900 World        2 Medium   1950      1950 45-49           45          5
#> # … with 909,710 more rows, and 1 more variable: PopFemale <dbl>

# tidy sex into a single column and drop id columns
get_wpp(indicator = "pop", pop_sex = "both", pop_age = "five",
        tidy_pop_sex = TRUE, drop_id_cols = TRUE)
#> Indicator(s) appears in more than one file.
#> 
#> Only downloading indicator(s) from file: Population1JanuaryByAge5GroupSex 
#> 
#> Need multiple get_wpp() calls to get indicators in different files. See ?wpp_indicators and ?find_indicators for more information on files.
#> # A tibble: 1,819,440 × 6
#>    Location Variant  Time AgeGrp Sex        Pop
#>    <chr>    <chr>   <dbl> <fct>  <chr>    <dbl>
#>  1 World    Medium   1950 0-4    Male   170451.
#>  2 World    Medium   1950 0-4    Female 163207.
#>  3 World    Medium   1950 5-9    Male   136460.
#>  4 World    Medium   1950 5-9    Female 130304.
#>  5 World    Medium   1950 10-14  Male   131621.
#>  6 World    Medium   1950 10-14  Female 125620.
#>  7 World    Medium   1950 15-19  Male   119914.
#>  8 World    Medium   1950 15-19  Female 116088.
#>  9 World    Medium   1950 20-24  Male   110160.
#> 10 World    Medium   1950 20-24  Female 108389.
#> # … with 1,819,430 more rows

# alternatively use indicator_file to select the desired version of population indicator(s)
get_wpp(indicator = c("PopTotal", "PopMale", "PopFemale"), indicator_file =  "TotalPopulationBySex")
#> # A tibble: 43,186 × 9
#>    LocID Location VarID Variant  Time MidPeriod  PopMale PopFemale PopTotal
#>    <dbl> <chr>    <dbl> <chr>   <dbl>     <dbl>    <dbl>     <dbl>    <dbl>
#>  1   900 World        2 Medium   1950     1950. 1245363.  1253959. 2499322.
#>  2   900 World        2 Medium   1951     1952. 1267473.  1275657. 2543130.
#>  3   900 World        2 Medium   1952     1952. 1291400.  1298871. 2590271.
#>  4   900 World        2 Medium   1953     1954. 1316659.  1323620. 2640279.
#>  5   900 World        2 Medium   1954     1954. 1342923.  1349056. 2691979.
#>  6   900 World        2 Medium   1955     1956. 1370392.  1375680. 2746072.
#>  7   900 World        2 Medium   1956     1956. 1398205.  1402798. 2801003.
#>  8   900 World        2 Medium   1957     1958. 1426956.  1430911. 2857867.
#>  9   900 World        2 Medium   1958     1958. 1456539.  1459569. 2916108.
#> 10   900 World        2 Medium   1959     1960. 1483989.  1486304. 2970292.
#> # … with 43,176 more rows

# clean column names
get_wpp(indicator = c("SRB", "NetMigrations", "PopGrowthRate"),
        clean_names = TRUE, drop_id_cols = TRUE)
#> # A tibble: 43,472 × 6
#>    location variant  time pop_growth_rate   srb net_migrations
#>    <chr>    <chr>   <dbl>           <dbl> <dbl>          <dbl>
#>  1 World    Medium   1950            1.73  105.              0
#>  2 World    Medium   1951            1.74  105.              0
#>  3 World    Medium   1952            1.93  105.              0
#>  4 World    Medium   1953            1.90  105.              0
#>  5 World    Medium   1954            1.98  105.              0
#>  6 World    Medium   1955            2.00  105.              0
#>  7 World    Medium   1956            1.96  105.              0
#>  8 World    Medium   1957            2.05  105.              0
#>  9 World    Medium   1958            1.98  105.              0
#> 10 World    Medium   1959            1.70  105.              0
#> # … with 43,462 more rows
# }